pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent_py"
version = "0.1.0"
description = "智能体生图应用 - 支持多Provider、并行批量生成"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "openpyxl>=3.1.0",
    "pandas>=2.1.0",
    "xlrd>=2.0.0",
    "pdfplumber>=0.10.0",
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "openai>=1.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "pypdf>=5.0.0",
    "sentence-transformers>=2.2.0",
    "Pillow>=10.0.0",
    "cryptography>=42.0.0",
    "minio>=7.2.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "aiosqlite",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "loguru>=0.7.0",
]

[project.scripts]
agent-py = "src.api.main:run"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"


